Refactor rng states component
